refactor piece selection handlers, rotate and flip handlers per feedback
